Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.10
python-decouple==3.8
django-cors-headers==4.3.0
Pillow>=10.0.0
celery==5.3.4
redis==5.0.1
PyPDF2==3.0.1
reportlab==4.0.7
pandas>=2.2.3
openpyxl==3.1.2
setuptools
fpdf2>=2.7.0
requests>=2.31.0
# AI Attendance System dependencies
# NOTE: dlib requires CMake + C++ build tools. On Windows install via:
#       pip install dlib-binary  (pre-built wheel, no CMake needed)
face_recognition==1.3.0
opencv-python==4.8.1.78
qrcode[pil]==7.4.2
# LLM: Uses OpenRouter API via 'requests' (already listed above)
